Departamen + Instructor + de todo @.@

+++ Universiti.BL/Models/Department.cs
@@ -11,7 +11,7 @@ namespace Universiti.BL.Models
         [Key]
         public int DepartmentID { get; set; }
         public string Name { get; set; }
-        public double Budget { get; set; }
+        public decimal Budget { get; set; }
         public DateTime StartDate { get; set; }
         
         [ForeignKey("Instructor")]
